Improve string handling, add run() and join(), and clean up module
Since strings are iterable but not arbitrary containers, more often than not they should be treated as atomic values and not as divisible objects. With this commit, Sq.from() treats strings as atomic values and the new Sq.fromString() supports building sequences of character strings. Similarly flatten() treats strings as non-iterable.

run() provides extensibility wiithout modifying the class or even coding anything different: Just invoke run() on a generator function taking its source iterable as the only argument. And join() comes in handy, even if it is trivially implemented with existing methods.