Add support for async concat() and zip(), restructure code to use class for Sq
Refactoring the implementation to use one superclass, Sq, and two subclasses, Sequence and AsyncSequence, makes the code much more readable and the layout of this rather large module clearer: After some helpers, there is the implementation of the sequence operators, all in pair of sync and async. Then come the three classes. One side effect of this change is that calling Sq() as a function does not work any more.