Rename @grr/multitasker to @grr/async and simplify package
The variant with three priorities has turned out to be a bit too brittle. The new version ditches priorities, which makes it easier to understand. I may re-add support for high priority execution but that wouldn't quite solve the problems I ran into when writing a file system walk with the previous version, i.e., multitasker. The test case for this version also is fs.walk, which I'm implementing next but with a very different API to its first version.