Add the new and actually clearly improved file system walk
Not only does it work, as evidenced by tests, its resource consumption is configurable. Out of the box, it uses no concurrency. But with, say, @gtrr/ async, it can also perform many small tasks concurrently.

Since tests are isolated in the top-level package and have their own directoy, there is no need for the verbose .test.js double extension. That practice ends with this commit.