Capture most of start.js functionality in @grr/loader/launch
The module exports a function that testts for presence of the custom module loader and spawns another process with that loader if not present. Conveniently, launch() only needs to be called once, at the end of the main module. Everything works out of the box with exception of the main function, which you should provide as an argument.