Separate validation from model; validate more thoroughly; clean up model
This commit is a major refactoring of @grr/model. While the separate and more thorough validation is an improvement (and caught a few bugs in the model), it still is too verbose and cumbersome. A more general schema validation library is called for.

The model clean up affects several properties, which now have more formal names. E.g., an attribute isn't an "instance" or hasn't an "tokens" anymore but rather is either an instance of the (named) "type" or "enum".  Similarly, an element  doesn't have "children" anymore but rather "content".  The type names now follow Java class naming conventions and the effective types for common enumerations have become the types period. "<custom-element>" and inter-element ">text<" have a new notation as well.

This commit also adds ARIA 1.2 attributes.