Add new @grr/schemata package
It's not quite complete, seeing that it still lacks support for Namespace() declaration and for checking keys as to belonging to one of these namespaces. Interestingly, this
is the first validation that requires propagating sets of data to the root, where we start a more targeted inspection after the first full inspection.