Fix OpenSSL configuration for self-certifying localhost
The previous configuration did result in a self-signed certificate. Yet when presented as a server's TLS certificate, it did not validate under itself as certificate authority. To make matters worse, the resulting error message "unable to verify first certificate" can be triggered by so many different faults that it is mostly useless.

The bottom line is that keyUsage must include both digitalSignature and keyCertSign (not keyEncipherment). Ironically, when dropping digitalSignature, the error message "unsupported certificate purpose" is on-point and helpful.

This change also removes the Netscape extension because it is officially deprecated.