Significantly enrich @grr/http
The package now has Client and Server classes to manage endpoints and an Exchange class to manage request/response interactions. Supporting modules include date parsing and pretty-printing of endpoints. The exports of the package have been updated accordingly. A few tests have been added as well and the basics of GET and HEAD requests work.