Rename Exchange to Context, improve data model, add critical functionality
Rename Exchange to Context to avoid introducing more lingo (Koa.js and others use Context, so no need to confuse everyone).

Add Message, Request, and Response classes to cleanly model messages having a set of headers and a separate body. The Request and Response subclasses have handwritten getters and setters for pseudo-headers that are valid for that message. Neither of the three classes has significant logic.

Rewrite functionality in Context not to use private message fields and to do the heavy lifting of built-in middleware, including for validating request paths, generating response bodies, redirecting requests, sending files as static assets, handling errors even if thrown in arbitrary code, and actually sending responses. Sending files is now just one instance of a readable stream providing the body, which integrates much better with other packages that handle streams. Node.js's sendFile() API is powerful but not extensible or interceptable. Streams are.

In short, Message, Request, and Response hold the state for the HTTP exchange whereas Context provides all the operations.